pyproject.toml
[project]
name = "discord-assistant"
version = "0.1.0"
description = "Always-on Discord AI assistant powered by MiniMax m2.5 via LangGraph"
requires-python = ">=3.13"
dependencies = [
    "discord.py>=2.4,<3",
    "langgraph>=0.4,<1",
    "langchain-openai>=0.3,<1",
    "langchain-core>=0.3,<1",
    "aiosqlite>=0.20,<1",
    "chromadb>=0.6,<1",
    "apscheduler>=3.10,<4",
    "playwright>=1.49,<2",
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "pyyaml>=6,<7",
    "python-dotenv>=1,<2",
    "structlog>=24,<26",
    "aiohttp>=3.9,<4",
    "firecrawl-py>=4.17,<5",
    "faster-whisper>=1.2.1",
    "yt-dlp>=2026.2.21",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "pytest-cov>=6,<7",
    "ruff>=0.9,<1",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
